Fix files structure
